fixed some compiler warnings in ppl6

+++ ppl6/src/ppl6_CThread.cpp
@@ -35,10 +35,10 @@
  * THE POSSIBILITY OF SUCH DAMAGE.
  *******************************************************************************/
 
+#include "prolog.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#include "prolog.h"
 
 #ifdef HAVE_STDARG_H
 	#include <stdarg.h>
@@ -1030,7 +1030,7 @@ int CThread::ThreadSetStackSize(size_t size)
 		#endif
 		THREADDATA *t=(THREADDATA *)threaddata;
 		if (size==0) size=PTHREAD_STACK_MIN;
-		if (size<PTHREAD_STACK_MIN) {
+		if (size<(size_t)PTHREAD_STACK_MIN) {
 			SetError(314,PTHREAD_STACK_MIN,"Stacksize muss mindestens %u Byte sein",PTHREAD_STACK_MIN);
 			return 0;
 		}
